+++ TabExpansionProfile.Invoke-Build.ps1
@@ -11,33 +11,28 @@
 	session from a PowerShell profile. This script should be placed to the
 	system path. It will be called on the first code completion.
 
-	This completion profile adds completers for Task and File arguments of
-	Build.ps1 and Invoke-Build.ps1
+	This script adds completers for Task and File arguments of Build.ps1 and
+	Invoke-Build.ps1
 #>
 
+# Common completer for File of Build.ps1 and Invoke-Build.ps1.
 $completeFile = {
 	param($commandName, $parameterName, $wordToComplete, $commandAst, $boundParameters)
-	$r = if ($wordToComplete) {
-		Get-ChildItem -Include "$wordToComplete*" -Name
+	Get-ChildItem -Directory -Name "$wordToComplete*"
+	if (!($boundParameters['Task'] -eq '**')) {
+		Get-ChildItem -File -Name "$wordToComplete*.ps1"
 	}
-	else {
-		Get-ChildItem -Include '*.build.ps1', '*.test.ps1' -Name
-	}
-	$r | .{process{ New-CompletionResult $_ }}
 }
 
+# Add completers to the option table
 $TabExpansionOptions.CustomArgumentCompleters += @{
 	'Build.ps1:Task' = {
 		param($commandName, $parameterName, $wordToComplete, $commandAst, $boundParameters)
-		$file = $boundParameters['File']
-		$all = Build.ps1 ?? -File $file
-		$all.Keys -like "$wordToComplete*" | .{process{ New-CompletionResult $_ }}
+		(Build.ps1 ?? -File ($boundParameters['File'])).Keys -like "$wordToComplete*"
 	}
 	'Invoke-Build.ps1:Task' = {
 		param($commandName, $parameterName, $wordToComplete, $commandAst, $boundParameters)
-		$file = $boundParameters['File']
-		$all = Invoke-Build.ps1 ?? -File $file
-		$all.Keys -like "$wordToComplete*" | .{process{ New-CompletionResult $_ }}
+		(Invoke-Build.ps1 ?? -File ($boundParameters['File'])).Keys -like "$wordToComplete*"
 	}
 	'Build.ps1:File' = $completeFile
 	'Invoke-Build.ps1:File' = $completeFile